+++ src/core/bx-remote.js
@@ -1,4 +1,4 @@
-KISSY.add('brix/core/bx-remote', function(S, appConfig, IO, Uri) {
+KISSY.add('brix/core/bx-remote', function(S, IO, Uri) {
 
     var exports = {
 
@@ -16,8 +16,10 @@ KISSY.add('brix/core/bx-remote', function(S, appConfig, IO, Uri) {
             else if (/^\.\//.test(remote)) {
                 var name = self.bxName
                 var mod = name.replace(/\/?$/, '') + remote.substr(1)
+                var family = mod.split('/')[0]
+                var packages = S.config('packages')[family]
 
-                if (appConfig.config('debug')) {
+                if (packages.debug) {
                     self.bxXhrRemote(mod, callback)
                 }
                 else {
@@ -67,7 +69,6 @@ KISSY.add('brix/core/bx-remote', function(S, appConfig, IO, Uri) {
     return exports
 }, {
     requires: [
-        'brix/app/config',
         'ajax',
         'uri'
     ]